Extract common volume viewport functionality to base class
