+++ packages/core/src/RenderingEngine/helpers/volumeNewImageEventDispatcher.ts
@@ -5,7 +5,7 @@ import {
 import { EventTypes } from '../../types';
 import { Events } from '../../enums';
 import { getRenderingEngine } from '../getRenderingEngine';
-import VolumeViewport from '../VolumeViewport';
+import BaseVolumeViewport from '../BaseVolumeViewport';
 
 // Keeping track of previous imageIndex for each viewportId
 type VolumeImageState = Record<string, number>;
@@ -35,9 +35,9 @@ function volumeNewImageEventDispatcher(
   const renderingEngine = getRenderingEngine(renderingEngineId);
   const viewport = renderingEngine.getViewport(viewportId);
 
-  if (!(viewport instanceof VolumeViewport)) {
+  if (!(viewport instanceof BaseVolumeViewport)) {
     throw new Error(
-      `volumeNewImageEventDispatcher: viewport is not a VolumeViewport`
+      `volumeNewImageEventDispatcher: viewport is not a BaseVolumeViewport`
     );
   }
