+++ packages/core/src/index.ts
@@ -9,6 +9,7 @@ import {
 } from './RenderingEngine';
 import RenderingEngine from './RenderingEngine';
 import VolumeViewport from './RenderingEngine/VolumeViewport';
+import BaseVolumeViewport from './RenderingEngine/BaseVolumeViewport';
 import StackViewport from './RenderingEngine/StackViewport';
 import Viewport from './RenderingEngine/Viewport';
 import eventTarget from './eventTarget';
@@ -62,6 +63,7 @@ export {
   //
   Settings,
   // Rendering Engine
+  BaseVolumeViewport,
   VolumeViewport,
   Viewport,
   StackViewport,
